# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pointme-Api",
        "description": "Documentation complète de l'API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register/admin": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "Enregistrer un nouvel administrateur",
                "description": "Crée un compte administrateur avec validation, upload de photo et hachage du mot de passe.",
                "operationId": "40f1c8b1f404c3f496a210b667119a66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "photo"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "photo": {
                                        "description": "Image de profil (jpg, jpeg, png)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Admin enregistré avec succès"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Jean"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Dupont"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jean.dupont@gmail.com"
                                                },
                                                "photo": {
                                                    "type": "string",
                                                    "example": "photos/jean_photo.jpg"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation échouée"
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "L'email est déjà utilisé"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur serveur"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "SQLSTATE[HY000] ..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "Connexion utilisateur (coach et stagiaire)",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/profileCoach": {
            "get": {
                "tags": [
                    "USER"
                ],
                "summary": "Profil de l'utilisateur connecté (coach)",
                "operationId": "bae6c007cbd0d03cd1c3a5cab40d7eed",
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profileStagiaire": {
            "get": {
                "tags": [
                    "USER"
                ],
                "summary": "Profil de l'utilisateur connecté (stagiaire)",
                "operationId": "aec70e4426b5b2b3f0dad735548605b0",
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "Déconnexion de l'utilisateur (coach et stagiaire)",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Entrer le token JWT précédé de \"Bearer \"",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}